feat(notification): 重写自定义属性示例(#2831)
