+++ packages/ui/toast/src/ToastAPI.tsx
@@ -6,7 +6,7 @@ import { ToastEventOptions } from './types'
 import { withDefaultProps } from '@hi-ui/react-utils'
 import { uuid } from '@hi-ui/use-id'
 
-export class ToastAPI<T extends ToastEventOptions> {
+export class ToastAPI<T = ToastEventOptions> {
   static defaultOptions = {
     prefixCls: _prefix,
   }
@@ -33,7 +33,7 @@ export class ToastAPI<T extends ToastEventOptions> {
     return `.${this.options.prefixCls}__portal__${this.id}`
   }
 
-  initManager(container: HTMLElement | undefined) {
+  initManager(container?: HTMLElement) {
     this.container = Container.getContainer(this.selector, document, container)
     this.toastManager = React.createRef<ToastManager>()
