+++ packages/ui/toast/src/ToastManager.tsx
@@ -69,9 +69,9 @@ export class ToastManager extends Component<ToastManagerProps, ToastManagerState
   }
 
   private getStyle = (
-    placement: ToastPlacement,
-    container?: HTMLElement | undefined,
-    zIndex?: number | undefined
+    placement: ToastPlacement = 'top',
+    container?: HTMLElement,
+    zIndex?: number
   ): React.CSSProperties => {
     let top: string | undefined
     let bottom: string | undefined
@@ -110,7 +110,7 @@ export class ToastManager extends Component<ToastManagerProps, ToastManagerState
     return (
       <div
         className={cx(`${prefixCls}-manager`, `${prefixCls}-manager--placement-${placement}`)}
-        style={this.getStyle(placement!, container, zIndex)}
+        style={this.getStyle(placement, container, zIndex)}
       >
         {queue.map((notice) => {
           if (As) return <As {...notice} />
@@ -144,8 +144,12 @@ export interface ToastManagerProps {
    * 放置 toast 的位置
    */
   placement?: ToastPlacement
-
-  container?: HTMLElement | undefined
-
-  zIndex?: number | undefined
+  /**
+   * 指定 portal 的容器
+   */
+  container?: HTMLElement
+  /**
+   * 自定义css展示层级
+   */
+  zIndex?: number
 }
